refactor: routing and navigation structure
