+++ web/package.json
@@ -13,7 +13,7 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"format": "prettier --write .",
 		"lint:fix": "eslint . --fix",
-		"lint": "prettier --check . && eslint . --cache",
+		"lint": "prettier --check . && eslint .",
 		"test:unit": "vitest --run --coverage",
 		"test": "npm run test:unit -- --run && npm run test:e2e",
 		"test:e2e": "playwright test",
