+++ web/src/lib/components/settings/extensions/utils.svelte.ts
@@ -13,7 +13,7 @@ function getAccordionValue(): ExtensionsBundleType[] | undefined {
 }
 
 function installExtensions(extensions: ExtensionsBundleType[]) {
-	const basePath = resolve('/(auth)/scope/[scope]/scope-based-settings/extensions', { scope: page.params.scope! });
+	const basePath = resolve('/(auth)/scope/[scope]/settings/extensions', { scope: page.params.scope! });
 	const url = new SvelteURL(basePath, window.location.origin);
 	extensions.forEach((extension) => url.searchParams.append(key, extension));
 	goto(url.pathname + url.search); // eslint-disable-line svelte/no-navigation-without-resolve
