+++ web/src/routes/(auth)/configuration/+layout.svelte
