+++ web/src/routes/(auth)/configuration/+page.svelte
