+++ web/src/routes/(auth)/configuration/single-sign-on/+page.svelte
@@ -1,0 +1,20 @@
+<script lang="ts">
+	import { resolve } from '$app/paths';
+	import { SingleSignOn } from '$lib/components/settings/single-sign-on';
+	import { m } from '$lib/paraglide/messages';
+	import { breadcrumbs } from '$lib/stores';
+
+	// Set breadcrumbs navigation
+	breadcrumbs.set([
+		{
+			title: m.settings(),
+			url: resolve('/(auth)/configuration'),
+		},
+		{
+			title: m.single_sign_on(),
+			url: resolve('/(auth)/configuration/single-sign-on'),
+		},
+	]);
+</script>
+
+<SingleSignOn />
