+++ web/src/routes/(auth)/machines/+page.svelte
@@ -1,0 +1,16 @@
+<script lang="ts">
+	import { resolve } from '$app/paths';
+	import Dashboard from '$lib/components/machines/dashboard/index.svelte';
+	import { m } from '$lib/paraglide/messages';
+	import { breadcrumbs } from '$lib/stores';
+
+	// Set breadcrumbs navigation
+	breadcrumbs.set([
+		{
+			title: m.machines(),
+			url: resolve('/(auth)/machines'),
+		},
+	]);
+</script>
+
+<Dashboard />
