--- web/src/routes/(auth)/scope/[scope]/machines/+page.svelte
@@ -1,17 +1,0 @@
-<script lang="ts">
-	import { resolve } from '$app/paths';
-	import { page } from '$app/state';
-	import Dashboard from '$lib/components/machines/dashboard/index.svelte';
-	import { m } from '$lib/paraglide/messages';
-	import { breadcrumbs } from '$lib/stores';
-
-	// Set breadcrumbs navigation
-	breadcrumbs.set([
-		{
-			title: m.machines(),
-			url: resolve('/(auth)/scope/[scope]/machines', { scope: page.params.scope! }),
-		},
-	]);
-</script>
-
-<Dashboard />
