+++ web/src/routes/(auth)/scope/[scope]/settings/+layout.svelte
