+++ web/src/routes/(auth)/scope/[scope]/settings/+page.svelte
